requirements.txt
matplotlib==3.7.1
numpy==1.24.1
pandas==1.5.3
PyYAML==6.0.2
scipy==1.14.1
statsmodels==0.13.5
tqdm==4.66.1
